# swagger_v1.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Job Portal API v1",
    "description": "Production-ready API backend for Flutter, Mobile, and Desktop apps",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost/MINDWAREINFOTECH",
      "description": "Local Development Server"
    }
  ],
  "tags": [
    {
      "name": "Authentication",
      "description": "User login and registration with JWT"
    },
    {
      "name": "Jobs",
      "description": "Public and protected job APIs"
    },
    {
      "name": "Profile",
      "description": "User profile management"
    },
    {
      "name": "Dashboard",
      "description": "User-specific dashboard stats"
    }
  ],
  "paths": {
    "/api/v1/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StandardResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/register-candidate": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Candidate Registration",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" },
                  "full_name": { "type": "string" },
                  "mobile": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" }
        }
      }
    },
    "/api/v1/jobs": {
      "get": {
        "tags": ["Jobs"],
        "summary": "List Jobs",
        "parameters": [
          { "name": "keyword", "in": "query", "schema": { "type": "string" } },
          { "name": "location", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Job list" }
        }
      }
    },
    "/api/v1/profile": {
      "get": {
        "tags": ["Profile"],
        "summary": "Get Profile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Profile data" }
        }
      },
      "put": {
        "tags": ["Profile"],
        "summary": "Update Profile",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "type": "object" }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated" }
        }
      }
    },
    "/api/v1/dashboard": {
      "get": {
        "tags": ["Dashboard"],
        "summary": "Get Dashboard Stats",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Dashboard data" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "StandardResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "boolean" },
          "message": { "type": "string" },
          "data": { "type": "object" },
          "errors": { "type": "object", "nullable": true }
        }
      }
    }
  }
}